refactor(decoder): split the 2,908-line src/decoder_managed.rs into src/decoder_managed/ modules
Pure move + visibility plumbing; no logic edits. src/decoder_managed.rs
becomes a directory module whose mod.rs is 38 lines: module docs, the
`#![deny(unsafe_code)]` inner attribute (lint level propagates to every
submodule — verified empirically by planting an unsafe block in
plane_convert.rs and watching the deny fire), the nine `mod` decls, and
the two explicit re-exports lib.rs consumes.

  decoder.rs       266  the struct + fields, new(), decode_frame /
                        decode_frame_inner (the sole cancel-token
                        install/uninstall site, kept adjacent), decode,
                        decode_full
  metadata.rs      278  resolved_matrix_for, build_image_info,
                        probe_info, set_native_gray, extract_gain_map,
                        animation_info, is_grid, grid_config
  frame_convert.rs 347  convert_to_image + convert_8bit / convert_16bit
                        + crop_image (Frame -> PixelBuffer driver)
  plane_convert.rs 672  ConvertCtx + the stateless YUV-plane kernels
  animation.rs     367  decode_animation, decode_anim_frame,
                        AnimationDecoder (+ AomAnimFrame)
  aom.rs           451  the whole aom-backend still/grid path
  sink.rs          311  decode_to_strip_converter, decode_to_sink,
                        decode_grid_to_sink
  grid.rs          275  decode_grid, stitch_tiles, stitch_tile_images,
                        decode_tile_row, stitch_tile_into_buffer + its
                        four stitch_tests (bodies unchanged)
  cicp_map.rs       89  rav1d <-> zenavif <-> yuv crate enum translation

Public API unchanged: mod.rs re-exports ManagedAvifDecoder and
AnimationDecoder BY NAME (no glob, so nothing can leak later); the two
pub structs, their 6 + 5 pub methods, and zero pub fields match
docs/public-api/zenavif.txt lines 180-185 / 282-288 exactly. Snapshots
were NOT regenerated (the generator enumerates unsafe-asm, unbuildable
on macOS aarch64) — non-drift is by inspection.

Cross-module helpers became pub(super) (struct fields, decode_frame,
convert_to_image, resolved_matrix_for, build_image_info,
extract_gain_map, decode_grid, stitch_tile_images, decode_full_aom,
convert_aom_to_image, ConvertCtx + kernels, the cicp_map fns). All six
#[allow(dead_code)], both #[cfg_attr(not(aom-backend), allow(dead_code))]
field markers, and the #[expect(clippy::too_many_arguments)] moved with
their items. The three aom cfgs inside the extracted impl/free fns are
now carried once by `#[cfg(feature = "aom-backend")] mod aom;` (12
attribute sites remain, was 14).

Also fixed four doc comments that a previous reorder had stranded on the
wrong item (prose only, no doctests involved): the grid-stitch paragraph
was on ConvertCtx (which this split would have moved to a different
file), probe_info's docs were on set_native_gray, the 8-bit monochrome
dispatch line was on convert_8bit_monochrome_gray, and the 8-bit planar
paragraph was on convert_8bit_identity. Each moved to the item it
describes. Left alone and reported: the AnimationDecoder rustdoc +
no_run doctest is attached to the cfg'd AomAnimFrame alias, so the
doctest only exists under aom-backend — fixing it would change
per-feature doctest collection, which is not a behavior-preserving move.

Verified: gauntlet check PASS on default/encode/aom/svt/tq/backends/
expert/autotune/twopass/allsafe (--all-features still fails on
rav1d's C .S sources, pre-existing on this box); nextest default
345/345 pass; nextest allsafe 586/587, the one failure being the known
12bpc product_aom_backend animations_decode_identically_across_backends
(imazen/zenav1-aom#8). Doctests 7 (default) / 8 (aom) — unchanged.
rustfmt clean for the new files (it also reflowed two over-long
Self::decode_frame calls the monolith carried).
